+++ routes.test.js
@@ -1,7 +1,7 @@
 const request = require('supertest');
 const GameServer = require('backend/js/GameServer');
 const GameRoom = require('backend/js/GameRoom');
-const Player = require('backend/js/Player');
+const Player = require('backend/js/GameRoom/Player');
 const { mockSend } = require('common/mocks');
 const { initSocket, destroySocket, TEST_PORT } = require('common/mocks/wsMocks');
 const pubSub = require('backend/helpers/pubSub');
