+++ src/helpers/routeHelpers.js
@@ -1,7 +1,7 @@
 const { GAME_TYPES } = require('common/constants');
 const { PLAY, EXECUTE_COMMANDS } = require('backend/topics');
 const GameServer = require('backend/js/GameServer');
-const Player = require('backend/js/Player');
+const Player = require('backend/js/GameRoom/Player');
 const pubSub = require('backend/helpers/pubSub');
 
 const multiGameExists = (gameId) => {
