+++ src/js/GameRoom/MessageManager.js
