+++ src/js/GameRoom/Player/__tests__/ServerBoard.test.js
@@ -1,10 +1,10 @@
-const ServerBoard = require('backend/js/ServerBoard');
 const { SEED_PIECES } = require('backend/constants');
 const { GET_PIECES, CLEAR_LINES } = require('backend/topics');
 const {
   getTestBoard, getTestPiece, getTestPieces, pubSubMock,
 } = require('common/mocks');
 const pubSub = require('backend/helpers/pubSub');
+const ServerBoard = require('../ServerBoard');
 
 describe('server - board tests', () => {
   let gameBoard;
