+++ src/js/GameRoom/PlayerManager.js
